+++ lib/cluster/ClusterOptions.ts
@@ -1,5 +1,13 @@
-import { lookup } from "dns";
+import { SrvRecord, resolveSrv, lookup } from "dns";
 import { NodeRole } from "./util";
+
+export type DNSResolveSrvFunction = (
+  hostname: string,
+  callback: (
+    err: NodeJS.ErrnoException | undefined,
+    records?: SrvRecord[]
+  ) => void
+) => void;
 
 export type DNSLookupFunction = (
   hostname: string,
@@ -87,6 +95,17 @@ export interface IClusterOptions {
   retryDelayOnTryAgain?: number;
 
   /**
+   * By default, this value is 0, which means when a `MOVED` error is received,
+   * the client will resend the command instantly to the node returned together with
+   * the `MOVED` error. However, sometimes it takes time for a cluster to become
+   * state stabilized after a failover, so adding a delay before resending can
+   * prevent a ping pong effect.
+   *
+   * @default 0
+   */
+  retryDelayOnMoved?: number;
+
+  /**
    * The milliseconds before a timeout occurs while refreshing
    * slots from the cluster.
    *
@@ -117,6 +136,23 @@ export interface IClusterOptions {
    * @default false
    */
   lazyConnect?: boolean;
+
+  /**
+   * Discover nodes using SRV records
+   *
+   * @default false
+   */
+  useSRVRecords?: boolean;
+
+  /**
+   * SRV records will be resolved via this function.
+   *
+   * You may provide a custom `resolveSrv` function when you want to customize
+   * the cache behavior of the default function.
+   *
+   * @default require('dns').resolveSrv
+   */
+  resolveSrv?: DNSResolveSrvFunction;
 
   /**
    * Hostnames will be resolved to IP addresses via this function.
@@ -159,11 +195,14 @@ export const DEFAULT_CLUSTER_OPTIONS: IClusterOptions = {
   enableReadyCheck: true,
   scaleReads: "master",
   maxRedirections: 16,
+  retryDelayOnMoved: 0,
   retryDelayOnFailover: 100,
   retryDelayOnClusterDown: 100,
   retryDelayOnTryAgain: 100,
   slotsRefreshTimeout: 1000,
   slotsRefreshInterval: 5000,
+  useSRVRecords: false,
+  resolveSrv: resolveSrv,
   dnsLookup: lookup,
   enableAutoPipelining: false,
   autoPipeliningIgnoredCommands: [],