+++ lib/connectors/StandaloneConnector.ts
@@ -1,5 +1,5 @@
 import { createConnection, TcpNetConnectOpts, IpcNetConnectOpts } from "net";
-import { connect as createTLSConnection, SecureContextOptions } from "tls";
+import { connect as createTLSConnection, ConnectionOptions } from "tls";
 import { CONNECTION_CLOSED_ERROR_MSG } from "../utils";
 import AbstractConnector, { ErrorEmitter } from "./AbstractConnector";
 import { NetStream } from "../types";
@@ -11,18 +11,21 @@ export function isIIpcConnectionOptions(
 }
 
 export interface ITcpConnectionOptions extends TcpNetConnectOpts {
-  tls?: SecureContextOptions;
+  tls?: ConnectionOptions;
 }
 
 export interface IIpcConnectionOptions extends IpcNetConnectOpts {
-  tls?: SecureContextOptions;
+  tls?: ConnectionOptions;
 }
 
+type IStandaloneConnectionOptions = (
+  | ITcpConnectionOptions
+  | IIpcConnectionOptions
+) & { disconnectTimeout: number };
+
 export default class StandaloneConnector extends AbstractConnector {
-  constructor(
-    protected options: ITcpConnectionOptions | IIpcConnectionOptions
-  ) {
-    super();
+  constructor(protected options: IStandaloneConnectionOptions) {
+    super(options.disconnectTimeout);
   }
 
   public connect(_: ErrorEmitter) {
@@ -75,6 +78,10 @@ export default class StandaloneConnector extends AbstractConnector {
           reject(err);
           return;
         }
+
+        this.stream.once("error", (err) => {
+          this.firstError = err;
+        });
 
         resolve(this.stream);
       });