+++ test/unit/connectors/connector.ts
@@ -20,7 +20,7 @@ describe("StandaloneConnector", () => {
     });
 
     it("ignore path when port is set and path is null", async () => {
-      const spy = sinon.stub(net, "createConnection");
+      const spy = sinon.spy(net, "createConnection");
       const connector = new StandaloneConnector({ port: 6379, path: null });
       await connector.connect(() => {});
       expect(spy.calledOnce).to.eql(true);
@@ -29,14 +29,19 @@ describe("StandaloneConnector", () => {
     });
 
     it("supports tls", async () => {
-      const spy = sinon.stub(tls, "connect");
+      const spy = sinon.spy(tls, "connect");
       const connector = new StandaloneConnector({
         port: 6379,
-        tls: { ca: "on" },
+        tls: { ca: "on", servername: "localhost", rejectUnauthorized: false },
       });
       await connector.connect(() => {});
       expect(spy.calledOnce).to.eql(true);
-      expect(spy.firstCall.args[0]).to.eql({ port: 6379, ca: "on" });
+      expect(spy.firstCall.args[0]).to.eql({
+        port: 6379,
+        ca: "on",
+        servername: "localhost",
+        rejectUnauthorized: false,
+      });
       connector.disconnect();
     });
   });